requirements.txt
torch>=1.10.0
torchvision>=0.11.1
scikit-image
pillow
numpy
tqdm
tensorboard
tensorboardX
face-alignment
opencv-python
matplotlib
timm>=0.6.0
pytorch-fid
torch-fidelity
clean-fid
